Rename error.rb => errors.rb
