+++ lib/dry/configurable/errors.rb
